GUIをWPFに移行/--mic_record_min_duration対応
