+++ src/cs-recognition-frontend/App.xaml.cs
@@ -1,0 +1,12 @@
+using System.Configuration;
+using System.Data;
+using System.Windows;
+
+namespace Haru.Kei;
+
+/// <summary>
+/// Interaction logic for App.xaml
+/// </summary>
+public partial class App : global::System.Windows.Application {
+}
+
