+++ src/cs-recognition-frontend/Views/MainWindow.xaml
@@ -1,0 +1,50 @@
+<Window
+    x:Class="Haru.Kei.Views.MainWindow"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    xmlns:local="clr-namespace:Haru.Kei.Views"
+    xmlns:wf="clr-namespace:System.Windows.Forms;assembly=System.Windows.Forms"
+    mc:Ignorable="d"
+    Title="ゆーかねすぴれこランチャー"
+    Height="480" Width="820"
+    Background="{DynamicResource {x:Static SystemColors.ControlBrushKey}}">
+    <Grid>
+        <Grid.RowDefinitions>
+            <RowDefinition Height="Auto" />
+            <RowDefinition Height="*" />
+            <RowDefinition Height="60" />
+        </Grid.RowDefinitions>
+        <Menu>
+            <MenuItem Header="ファイル(_F)">
+                <MenuItem Header="バッチファイルを作成" Command="{Binding CreateBatchommand}"/>
+                <Separator />
+                <MenuItem Header="マイクテスト" Command="{Binding MicTestCommand}"/>
+                <MenuItem Header="環境音測定" Command="{Binding AmbientTestCommand}"/>
+                <Separator />
+                <MenuItem Header="終了(_E)" Command="{Binding CloseCommand}"/>
+            </MenuItem>
+            <MenuItem Header="プリセット(_P)">
+                <MenuItem Header="音声認識にwhisperを使用" Command="{Binding ConnectWhisperCommand}"/>
+                <MenuItem Header="音声認識にgoogleを使用"  Command="{Binding ConnectGoogleCommand}"/>
+                <Separator />
+                <MenuItem Header="ゆかりねっとと連携"  Command="{Binding ConnectYukarinetteCommand}"/>
+                <MenuItem Header="ゆかコネNEOと連携"  Command="{Binding ConnectYukaConeCommand}"/>
+            </MenuItem>
+        </Menu>
+        <WindowsFormsHost Grid.Row="1">
+            <wf:PropertyGrid x:Name="propertyGrid" />
+        </WindowsFormsHost>
+        <Button
+            Grid.Row="2"
+            Width="92" Height="45"
+            HorizontalAlignment="Right"
+            VerticalAlignment="Bottom"
+            Margin="5"
+            Content="起動"
+            Command="{Binding ExecCommand}"
+            />
+
+    </Grid>
+</Window>
