+++ src/cs-recognition-frontend/recognize-frontend.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.8.34330.188
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "recognize-frontend", "recognize-frontend.csproj", "{9B293B5D-4C82-42F3-BF43-E72F0823AC45}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "recognize-frontend", "recognize-frontend.csproj", "{F99261A0-3EE0-3FB8-42B9-94CA640DC76D}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -11,10 +11,10 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{9B293B5D-4C82-42F3-BF43-E72F0823AC45}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9B293B5D-4C82-42F3-BF43-E72F0823AC45}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9B293B5D-4C82-42F3-BF43-E72F0823AC45}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9B293B5D-4C82-42F3-BF43-E72F0823AC45}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F99261A0-3EE0-3FB8-42B9-94CA640DC76D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F99261A0-3EE0-3FB8-42B9-94CA640DC76D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F99261A0-3EE0-3FB8-42B9-94CA640DC76D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F99261A0-3EE0-3FB8-42B9-94CA640DC76D}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
