--- src/cs-recognition-frontend/src/Program.cs
@@ -1,15 +1,0 @@
-using System;
-using System.Windows.Forms;
-
-
-namespace Haru.Kei {
-	internal static class Program {
-		/// <summary>
-		///  The main entry point for the application.
-		/// </summary>
-		[STAThread]
-		static void Main() {
-			Application.Run(new Form1());
-		}
-	}
-}
